backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
firebase-admin==6.4.0
google-cloud-aiplatform==1.42.1
pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.1
